# api.yaml
openapi: 3.0.0
info:
  description: Документация к нашему REST чудовищу
  version: 1.0.0
  title: Drello backend API
...
servers:
  - url: http://localhost:8080
security:
  - app_id: []
...
paths:
  /settings:
    post:
      tags:
        - settings
      summary: Регистрируем пользователя в нашем сервисе
      description: ...
      requestBody:
        description: Данные, необходимые для регистрации
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Board/properties/admins/items"
            example:
              name: Vasya
              surname: Pupkin
              nickname: Legolas1
              email: VasyaPupkin@gmail.com
              password: qwerty1234
      responses:
        "200":
          description: Успешная регистрация
          headers:
            session_id:
              schema:
                type: string
              description: Наша авторизационная кука (пример -
                4d756d94-b9d8-48dd-a4fe-cdc14d251ec2).
        "303":
          description: Постучались на данный endpoint с выставленной кукой
        "400":
          description: Тело запроса не соответствует модели пользователя
        "409":
          description: Пользователь с такими регистрационными данными уже существует
    get:
      tags:
        - settings
      summary: Получаем все данные о пользователе (у нас сейчас приходят одни и те же
        данные, что из /profile/{..}
      description: ...
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Валидный запрос данных пользователя
          content:
            application/json:
              schema:
                - $ref: "#/components/schemas/Board/properties/admins/items"
              example:
                id: 110323
                name: Vasya
                surname: Pupkin
                nickname: Legolas1
                email: VasyaPupkin@gmail.com
                avatar: Обсуждаем
        "401":
          description: В запросе отсутствует кука
        "500":
          description: Внутренная ошибка при маршалинге найденного пользователя
    put:
      tags:
        - settings
      summary: Обновляем данные пользователя
      description: ...
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      requestBody:
        description: Данные, необходимые для обновления (все завернуто в мультипарт-форму)
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                newName:
                  type: string
                newSurname:
                  type: string
                newNickname:
                  type: string
                newEmail:
                  type: string
                oldPassword:
                  type: password
                newPassword:
                  type: password
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Успешно обновили
        "400":
          description: Невалидное тело запроса с информацией для обновления
        "401":
          description: В запросе отсутствует кука
        "412":
          description: Неверный пароль
    delete:
      tags:
        - settings
      summary: Удаляем нашего пользователя
      description: ...
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Успешное удаление
        "401":
          description: В запросе отсутствует кука
        "500":
          description: Ошибка на сервере при удалении
...
  "/profile/{id_or_nickname}":
    get:
      tags:
        - profile
      summary: Запрос к публичным данным пользователя по идентификатору в сервисе
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: id пользователя, публичные данные которого хотим просмотреть
          example: 1106234
      responses:
        "200":
          description: Корректный запрос, существующий пользователь, данные в теле ответа
          content:
            application/json:
              schema:
                - $ref: "#/components/schemas/Board/properties/admins/items"
              example:
                name: Vasya
                surname: Pupkin
                nickname: Legolas1
                email: VasyaPupkin@gmail.com
                avatar: Обсуждаем
        "404":
          description: Пользователь не найден
        "500":
          description: Внутренная ошибка при маршалинге найденного пользователя
  /boards:
    post:
      tags:
        - boards
      summary: Создаем доску
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      requestBody:
        description: Данные, необходимые для создания доски
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - title
              properties:
                id:
                  type: integer
                title:
                  type: string
                columns:
                  type: array
                  items:
                    $ref: "#/components/schemas/Board/properties/columns/items"
                admins:
                  type: array
                  items:
                    $ref: "#/components/schemas/Board/properties/admins/items"
                members:
                  type: array
                  items:
                    $ref: "#/components/schemas/Board/properties/admins/items"
            example:
              title: MyFirstBoard
      responses:
        "200":
          description: Успешно получили данные по доске
          content:
            application/json:
              schema:
                $ref: "#/paths/~1boards/post/requestBody/content/application~1json/schema"
              example:
                id: 10
                title: SuperBoard
                admins:
                  - id: 110
                    name: Sasha
                    surname: Upkin
                    nickname: Vova1
                    email: SashaUpkin@gmail.com
                    avatar: http://localhost:5757/img/avatar/Vova1.png
        "400":
          description: Невалидные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "500":
          description: Внутренняя ошибка сервера
    get:
      tags:
        - boards
      summary: Получаем все доски, к которым пользователь имеет отношение
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Успешно получили список досок (пустой список в том числе)
          content:
            application/json:
              schema:
                $ref: "#/paths/~1boards/post/requestBody/content/application~1json/schema"
              example:
                member:
                  - id: 14
                    title: MemberBoard
                    members:
                      - id: 110323
                        name: Vasya
                        surname: Pupkin
                        nickname: Legolas1
                        email: VasyaPupkin@gmail.com
                        avatar: http://localhost:5757/img/avatar/Legolas1.png
                    admins:
                      - id: 110
                        name: Sasha
                        surname: Upkin
                        nickname: Vova1
                        email: SashaUpkin@gmail.com
                        avatar: http://localhost:5757/img/avatar/Vova1.png
                admin:
                  - id: 35
                    title: AdminBoard
                    members:
                      - id: 110323
                        name: Sasha
                        surname: Upkin
                        nickname: Vova1
                        email: VasyaPupkin@gmail.com
                        avatar: http://localhost:5757/img/avatar/Vova1.png
                    admins:
                      - id: 110
                        name: Vasya
                        surname: Ppkin
                        nickname: Legolas1
                        email: SashaUpkin@gmail.com
                        avatar: http://localhost:5757/img/avatar/Legolas1.png
        "401":
          description: В запросе отсутствует кука
        "500":
          description: Внутренняя ошибка сервера
...
components:
...
  securitySchemes:
    app_id:
      type: apiKey
      description: API key to authorize requests. If you don't have an OpenWeatherMap
        API key, use `fd4698c940c6d1da602a70ac34f0b147`.
      name: appid
      in: query